+++ Sources/GravatarUI/Base/AssociatedObject.swift
