+++ Sources/GravatarUI/Base/Box.swift
