+++ Sources/GravatarUI/Base/SimpleCounter.swift
