+++ Sources/GravatarUI/Options/ActivityIndicatorType.swift
