+++ Sources/TestHelpers/ImageHelper.swift
@@ -1,24 +1,24 @@
 import UIKit
 
-class ImageHelper {
+public class ImageHelper {
     private init() {}
 
     static let testResourcesDir = "Gravatar_Gravatar-Tests.bundle/ResourceFiles/"
 
-    static var testImage: UIImage {
+    public static var testImage: UIImage {
         image(named: "test", type: "png")!
     }
 
-    static var testImageData: Data {
+    public static var testImageData: Data {
         dataFromImage(named: "test", type: "png")!
     }
 
-    static var placeholderImage: UIImage {
+    public static var placeholderImage: UIImage {
         image(named: "placeholder", type: "png")!
     }
 
-    static func dataFromImage(named: String, type: String) -> Data? {
-        guard let url = Bundle.testsBundle.url(forResource: named, withExtension: type) else {
+    public static func dataFromImage(named: String, type: String) -> Data? {
+        guard let url = Bundle.module.url(forResource: named, withExtension: type) else {
             return nil
         }
         var data: Data? = nil
@@ -28,8 +28,8 @@ class ImageHelper {
         return data
     }
 
-    static func image(named: String, type: String) -> UIImage? {
-        guard let path = Bundle.testsBundle.path(forResource: named, ofType: type) else {
+    public static func image(named: String, type: String) -> UIImage? {
+        guard let path = Bundle.module.path(forResource: named, ofType: type) else {
             return nil
         }
         return UIImage(contentsOfFile: path)
