+++ Sources/TestHelpers/TestImageCache.swift
@@ -2,11 +2,13 @@ import Foundation
 import Gravatar
 import UIKit
 
-class TestImageCache: ImageCaching {
+public class TestImageCache: ImageCaching {
     var dict: [String: UIImage] = [:]
-    var getImageCallCount = 0
-    var setImageCallsCount = 0
+    public var getImageCallCount = 0
+    public var setImageCallsCount = 0
 
+    public init() { }
+    
     public func setImage(_ image: UIImage, forKey key: String) {
         setImageCallsCount += 1
         dict[key] = image
