+++ Tests/GravatarTests/TestURLSession.swift
@@ -1,5 +1,6 @@
 import Foundation
 import Gravatar
+import TestHelpers
 import XCTest
 
 enum TestDataTaskFailReason: Equatable {
